统一es返回得分  --2023-08-14

+++ dubbo-api/src/main/java/com/resume/dubbo/domian/Resume.java
@@ -137,7 +137,7 @@ public class Resume implements Serializable {
     private String img;
 
     @TableField(exist = false)
-    private double score;
+    private Float score;
 
     @TableField(exist = false)
     private List<String> tags_good = new ArrayList<>();
